requirements.txt
asgiref==3.8.1
Django==4.2.14
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==2.2.0
django-query-counter==0.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.7
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-environ==0.4.54
pytz==2024.1
PyYAML==6.0.2
setuptools==74.1.2
sqlparse==0.5.1
tabulate==0.9.0
tzdata==2024.1
uritemplate==4.1.1
